+++ backend/canisters/community/impl/src/updates/register_proposal_vote_v2.rs
@@ -3,7 +3,6 @@ use canister_api_macros::update;
 use canister_tracing_macros::trace;
 use community_canister::register_proposal_vote_v2::*;
 use types::OCResult;
-use utils::migrated_user_ids::MigratedUserIds;
 
 #[update(msgpack = true)]
 #[trace]
@@ -25,7 +24,7 @@ fn register_proposal_vote_impl(args: Args, state: &mut RuntimeState) -> OCResult
         args.message_index,
         args.adopt,
         now,
-        &MigratedUserIds::default(),
+        &state.data.migrated_user_ids,
     )?;
 
     channel.chat.members.register_proposal_vote(&user_id, args.message_index, now);