+++ backend/canisters/user/api/src/updates/mod.rs
@@ -10,6 +10,8 @@ pub mod c2c_bot_send_message;
 pub mod c2c_charge_user_account;
 pub mod c2c_community_canister;
 pub mod c2c_community_canister_v2;
+pub mod c2c_export_user;
+pub mod c2c_export_user_stable_memory;
 pub mod c2c_game_chit;
 pub mod c2c_grant_super_admin;
 pub mod c2c_group_canister;